Resonanze update
The envelope is now only a generator: you need a multiplier/VCA to apply
it to the synthesizer.

preparing for a digipotfilter component.
Here all the filter stuff will go in.
the filter can then automate to an envelope or modulation track

fake antilog

Resonance is now controllable with a digipot.
It ranges from 0 to very much.
No self oscillation yet with the state variable filter though...